+++ edgy/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-__version__ = "0.32.1"
+__version__ = "0.32.2"
 from typing import TYPE_CHECKING
 
 from ._monkay import Instance, create_monkay